pyproject.toml
[project]
name = "dotsOCR"
version = "0.1.0"
description = "Impementing dotOCR with Fronte"
readme = "README.md"
requires-python = "==3.13.*"


dependencies = [
    "openai>=1.5.0",
    "pillow>=10.4.0",
    "pymupdf>=1.24.0",
    "pydantic-settings",
    "streamlit>=1.41.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "mypy>=1.10.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.4.0",
    "ruff>=0.15.0",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dotsOCR"]

